Delete legacy_apihandler_auth.go, relocate types to mptypes.go
Auth peer routes (/auth/peer, /auth/distrib) were MP-only — tdns-authv2
has no peers without MP. Route registration removed from apirouters.go.
Types (AuthPeerPost, AuthPeerResponse, AuthDistribPost) relocated to
mptypes.go since they're referenced by CLI code.

Build verified: all tdns + tdns-mp binaries compile.
