+++ v2/apirouters.go
@@ -108,10 +108,8 @@ func (conf *Config) SetupAPIRouter(ctx context.Context) (*mux.Router, error) {
 		sr.HandleFunc("/delegation", APIdelegation(conf.Internal.DelegationSyncQ)).Methods("POST")
 	}
 
-	if Globals.App.Type == AppTypeAuth {
-		sr.HandleFunc("/auth/peer", APIauthPeer(conf)).Methods("POST")
-		sr.HandleFunc("/auth/distrib", APIauthDistrib(conf)).Methods("POST")
-	}
+	// Auth peer routes removed — peer management is MP-only.
+	// For tdns-mp signer, routes are registered via SetupMPSignerRoutes.
 
 	if Globals.App.Type == AppTypeAgent {
 		sr.HandleFunc("/agent", conf.APIagent(conf.Internal.RefreshZoneCh, kdb)).Methods("POST")
